% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{CalculateBootCI}
\alias{CalculateBootCI}
\title{Calculate column-wise confidence intervals from a numeric matrix}
\usage{
CalculateBootCI(cts, Times = 500, Cores = 1, CI = 0.95)
}
\arguments{
\item{cts}{a count matrix}

\item{Times}{number of resamplings}

\item{Cores}{numbers of mclapply workers}

\item{CI}{the confidence interval, default is 95\%}
}
\description{
Function to calculate confidence interval from a numeric matrix on a column-wise basis.
Code is adapted and basically identical to the Utilities.R script from the Bioconductor package ChIPSeeker.
Main intend is to serve as helper function for the PlotProfiles() function.
}
\details{

}
\examples{
cts <- sapply(seq(1,100), function(x) rnorm(100, 1))
ci <- CalculateBootCI(cts = cts, Times = 500, Cores = 8, CI = 0.95)

}
\author{
Alexander Toenges
}
