% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggMAplot.R
\name{ggMAplot}
\alias{ggMAplot}
\title{A ggplot wrapper to produce MAplots.}
\usage{
ggMAplot(
  xval,
  yval,
  pval = NULL,
  xval.thresh = NULL,
  yval.thresh = 0,
  pval.thresh = 0.05,
  basic.col = "grey50",
  col.up = "firebrick",
  col.down = "darkblue",
  Up = "Up",
  Down = "Down",
  NonSig = "NonSig",
  point.size = 1,
  point.alpha = 0.75,
  xlab = "baseMean",
  ylab = "logFC",
  xlim = NULL,
  ylim = NULL,
  title = waiver(),
  subtitle = waiver(),
  x.ablines = NULL,
  x.ablines.col = "black",
  x.ablines.lty = "dashed",
  y.ablines = NULL,
  y.ablines.col = "black",
  y.ablines.lty = "dashed",
  quantiles.x = c(0, 1),
  quantiles.y = c(0.001, 0.999),
  loess.span = NULL,
  loess.color = "black",
  loess.lty = "solid",
  preset = c("maplot", "volcano"),
  no.legend = FALSE
)
}
\arguments{
\item{xval}{a vector with values for the x-axis, usually average expression in MAs and 
fold change in Volcanos}

\item{yval}{a vector with values for the y-axis, usually fold change in MAs and -log10(pvalue) 
in Volcanos}

\item{pval}{a vector with pvalue-like values. Points below the \code{pval.thresh} will be 
highlighted}

\item{xval.thresh}{threshold for x-axis values, see details}

\item{yval.thresh}{threshold for y-axis values, see details}

\item{pval.thresh}{threshold for pvalues, see details}

\item{basic.col}{the color for the data points}

\item{col.up}{use this color for points classified as Up, see details}

\item{col.down}{use this color for points classified as Down, see details}

\item{Up}{legend label for points being "Up"}

\item{Down}{legend label for points being "Down"}

\item{NonSig}{legend label for points being "NonSig"}

\item{point.size}{size of data points}

\item{point.alpha}{point opacity}

\item{xlab}{x label}

\item{ylab}{y label}

\item{xlim}{xaxis limits, by default full data range}

\item{ylim}{yaxis limits, by default the 0.001th and 0.999th quantile of the data range, see details}

\item{title}{main title text}

\item{subtitle}{subtitle text}

\item{x.ablines}{a vector with numeric values, will use to draw vertical ablines}

\item{x.ablines.col}{color of x.ablines}

\item{x.ablines.lty}{line type of x.ablines}

\item{y.ablines}{a vector with numeric values, will use to draw horizontal ablines}

\item{y.ablines.col}{color of y.ablines}

\item{y.ablines.lty}{line type of y.ablines}

\item{quantiles.x}{use these quantiles of the xaxis data range to determine automatic axis limits, 
see details}

\item{quantiles.y}{use these quantiles of the yaxis data range to determine automatic axis limits, 
see details}

\item{loess.span}{if not NULL then use this value (e.g. 0.2) as span to fit a loess regression for
yval~xval and add to the plot}

\item{loess.color}{color of loess.span regression line}

\item{preset}{the type of plot, "maplot" by default, or "volcano" to make a volcano plot, 
see examples}

\item{no.legend}{logical, whether to remove the legend}
}
\description{
A ggplot wrapper for MA-plots and Volcanos
}
\details{
=> towards xval.thresh/yval.threshold/pval.threshold
We assume a typical differential expression results table, so with a fold change, 
an average expression value and some kind of p-value.
We can trigger color-highlighting of significant genes iF a pvalue is provided to the function. 
For MAplots points below this pval.thresh will then be classified as "Up" if yval > yval.thresh 
which is usually the fold change in MA-plots, or "Down" if < yval.thresh, or "NonSig" 
if > pval.thresh. xval.thresh is turned off by default for MAplots but if a threshold is provided
then only points with xval > xval.thresh will be considered for the classification.
In \code{preset="volcano"} the yval.thresh is ignored as the yaxis is simply -log10(pval),
therefore one should use \code{pval.thresh} for filtering.

=> towards quantiles.x/y
By default in MAplot mode the yaxis (fold changes) is automatically scaled to avoid overly wide
limits due to outliers. The 0.001th and 0.999th quantile of the yvalues is used to set the limits.
Points beyond these limits will be trimmed back to these limits and displayed as trianged rather
than dots. The user can set custom quantiles or simply use an explicit \code{xlim} or \code{ylim}
value to manually set axis limits. The automatic axis limit setting can be turned off by setting
\code{quantiles.x} or \code{quantiles.y} to NULL or \code{c(0,1)}.
}
\examples{
dds <- DESeq2::DESeq(DESeq2::makeExampleDESeqDataSet(5000,10))
res <- DESeq2::results(dds) \%>\% data.frame \%>\% na.omit
res$baseMean<-log2(res$baseMean+1)
theme_set(theme_classic(base_size = 12.5))

# a standard MA-plot (pvalue instead of padj simply because then we get more genes in color)
ggMAplot(xval = res$baseMean, yval = res$log2FoldChange, pval = res$pvalue,
title = "DE results", subtitle = "most basic type of MA-plot")

# with an abline at abs(log2(1.5)) and at 0:
ggMAplot(xval = res$baseMean, yval = res$log2FoldChange, pval = res$pvalue,
y.ablines = c(log2(1.5),-log2(1.5)), title = "DE results", subtitle = "with ablines")

# as Volcano:
ggMAplot(xval = res$log2FoldChange, yval = -log10(res$pvalue), pval = res$pvalue,
title = "DE results", subtitle = "Volcano plot", preset = "volcano")

# with modified y-limits (values beyond axis limits appear as trianges) from 0 to 2:
ggMAplot(xval = res$log2FoldChange, yval = -log10(res$pvalue), pval = res$pvalue,
title = "DE results", subtitle = "trimmed", preset = "volcano", ylim = c(0, 2))

}
\author{
Alexander Toenges
}
